requirements.txt
pytz
pandas
yfinance
requests
stocksymbol
python-binance
pandas-market-calendars
polygon
